move DocBlockManipulator to DocBlockUpdater
